+++ fennel/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 ## [1.1.4] - 2024-03-28
-- Add last operator to the client.
+- Add latest operator to the client.
 
 ## [1.1.3] - 2024-03-25
 - Add `erase_key` to Dataset to give user ability to erase certain entitiy from dataset